Change `pull_interval` to float value.

+++ lib/fluent/plugin/in_gcloud_pubsub.rb
@@ -12,7 +12,7 @@ module Fluent
     config_param :key,                :string,  :default => nil
     config_param :topic,              :string,  :default => nil
     config_param :subscription,       :string
-    config_param :pull_interval,      :integer, :default => 5
+    config_param :pull_interval,      :float,   :default => 5.0
     config_param :max_messages,       :integer, :default => 100
     config_param :return_immediately, :bool,    :default => true
     config_param :format,             :string,  :default => 'json'
